[Refactor] (Build): 将DEMO构建变体重命名为BOARDTEST
重构构建系统和相关代码，将DEMO变体更名为BOARDTEST以更准确反映其用途
更新CMake配置、预设文件和源代码中的相关定义
同步修改目录结构和文件命名

该变更使构建变体命名更加清晰明确
