+++ CMakeLists.txt
@@ -18,7 +18,7 @@ if(BUILD_VARIANT STREQUAL "MASTER")
 project(MasterFirmware VERSION 3.0)
 elseif(BUILD_VARIANT STREQUAL "SLAVE")
 project(SlaveFirmware VERSION 3.0)
-elseif(BUILD_VARIANT STREQUAL "DEMO")
+elseif(BUILD_VARIANT STREQUAL "BOARDTEST")
 project(BoardTestFirmware VERSION 3.0)
 else()
   message(FATAL_ERROR "Unknown BUILD_VARIANT value: ${BUILD_VARIANT}")
