+++ CMakePresets.json
@@ -7,8 +7,8 @@
     },
     "configurePresets": [
         {
-            "name": "master",
-            "displayName": "Master Edition",
+            "name": "Master",
+            "displayName": "Master",
             "toolchainFile": "${sourceDir}/CMake/arm-none-eabi-gcc.cmake",
             "generator": "Ninja",
             "binaryDir": "${sourceDir}/build",
@@ -19,8 +19,8 @@
             }
         },
         {
-            "name": "slave",
-            "displayName": "Slave Edition",
+            "name": "Slave",
+            "displayName": "Slave",
             "toolchainFile": "${sourceDir}/CMake/arm-none-eabi-gcc.cmake",
             "generator": "Ninja",
             "binaryDir": "${sourceDir}/build",
@@ -31,15 +31,15 @@
             }
         },
         {
-            "name": "demo",
-            "displayName": "Demo Edition",
+            "name": "BoardTest",
+            "displayName": "BoardTest",
             "toolchainFile": "${sourceDir}/CMake/arm-none-eabi-gcc.cmake",
             "generator": "Ninja",
             "binaryDir": "${sourceDir}/build",
             "cacheVariables": {
                 "CMAKE_EXPORT_COMPILE_COMMANDS": "true",
                 "CMAKE_BUILD_TYPE": "Debug",
-                "BUILD_VARIANT": "DEMO"
+                "BUILD_VARIANT": "BOARDTEST"
             }
         }
     ]
