+++ Source/Core/CMakeLists.txt
@@ -14,7 +14,7 @@ elseif(BUILD_VARIANT STREQUAL "SLAVE")
             ./slave/src/peripherals.cpp)
   target_include_directories(${EXECUTABLE_NAME} PUBLIC . ./slave/inc)
 
-elseif(BUILD_VARIANT STREQUAL "DEMO")
-  target_sources(${EXECUTABLE_NAME} PRIVATE ${COMMON_SOURCES} ./demo/demo.cpp ./demo/peripherals.cpp)
-  target_include_directories(${EXECUTABLE_NAME} PUBLIC . ./demo)
+elseif(BUILD_VARIANT STREQUAL "BOARDTEST")
+  target_sources(${EXECUTABLE_NAME} PRIVATE ${COMMON_SOURCES} ./BoardTest/board_test.cpp ./BoardTest/peripherals.cpp)
+  target_include_directories(${EXECUTABLE_NAME} PUBLIC . ./BoardTest)
 endif()
